Do not use the helper module or lodash/underscore.
There was very little to be gained from using lodash/underscore in the plugin
and in the test suite. The code of the plugin is meant to run in Node, and we've
dropped support of anything prior to Node 4. So having a library handle edge
cases on this and that browser is not useful.